Ensure line endings of scripts work when coming from Windows
If we are running the Docker image build from Windows, it's possible that the line endings will be changed from UNIX style to DOS.  This messes things up for our shell scripts we use as entry points in the Docker images.  We need to run the dos2unix conversion on our scripts when they are loaded, just to be safe.